+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- Added command line argument --list-modules which will list modules
+  and exit. It can be followed by "essential" or "optional" to only
+  list those modules.
+
 ## [0.12.0] - 2016-12-08
 ### Added
 - Added screenshots to Readme.
@@ -163,6 +169,7 @@ All notable changes to this project will be documented in this file.
 - Fixed issue that could allow someone to start a module setup process
   without being logged in to Plinth.
 
+[Unreleased]: https://github.com/freedombox/Plinth/compare/v0.12.0...HEAD
 [0.12.0]: https://github.com/freedombox/Plinth/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/freedombox/Plinth/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/freedombox/Plinth/compare/v0.9.4...v0.10.0